backend/requirements.txt
flask
flask-cors
groq
python-dotenv
pypdf
pyphen
google-genai
requests
pinecone
gdown
onnxruntime
onnxruntime-gpu
gunicorn
opencv-python
scipy
numpy
Pillow